+++ hw02/src/main/java/ru/sberbank/homework/gavrilov/MyCalculator.java
@@ -1,13 +1,15 @@
 package ru.sberbank.homework.gavrilov;
 
 import ru.sberbank.homework.common.Calculator;
-import ru.sberbank.homework.gavrilov.utils.Validator;
+import ru.sberbank.homework.gavrilov.utils.Helper;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.text.DecimalFormat;
 import java.text.DecimalFormatSymbols;
 import java.util.Locale;
+
+import static ru.sberbank.homework.gavrilov.utils.Parser.parseNumber;
 
 public class MyCalculator implements Calculator {
 
@@ -21,6 +23,11 @@ public class MyCalculator implements Calculator {
      */
     private BigDecimal resultBigDec;
 
+    private final Helper helper = new Helper();
+
+    private final DecimalFormatSymbols otherSymbols = new DecimalFormatSymbols(Locale.US);
+    private final DecimalFormat decimalFormat = new DecimalFormat("#.##", otherSymbols);
+
     /**
      * Processes user input.
      * Commands examples:
@@ -32,27 +39,26 @@ public class MyCalculator implements Calculator {
      */
     @Override
     public String calculate(String userInput) {
-        Validator validator = new Validator();
         BigDecimal first = null;
         BigDecimal second = null;
-        if (!validator.isValidCommand(userInput)) {
+        if (!helper.isValidCommand(userInput)) {
             resultBigDec = null;
             return WRONG_LINE;
         }
 
         try {
-            if (validator.getUsersCommand().length == 2) {
+            if (helper.getUsersCommand().length == 2) {
                 if (resultBigDec == null) {
                     throw new NumberFormatException(WRONG_LINE);
                 }
                 first = resultBigDec;
-                second = validator.parseNumber(validator.getUsersCommand()[1]);
+                second = parseNumber(helper.getUsersCommand()[1]);
             }
-            if (validator.getUsersCommand().length == 3) {
-                first = validator.parseNumber(validator.getUsersCommand()[0]);
-                second = validator.parseNumber(validator.getUsersCommand()[2]);
+            if (helper.getUsersCommand().length == 3) {
+                first = parseNumber(helper.getUsersCommand()[0]);
+                second = parseNumber(helper.getUsersCommand()[2]);
             }
-            resultBigDec = validator.execute().calculating(first, second);
+            resultBigDec = helper.execute().calculating(first, second);
         } catch (ArithmeticException | NumberFormatException exp) {
             resultBigDec = null;
             return exp.getMessage();
@@ -67,8 +73,6 @@ public class MyCalculator implements Calculator {
      * @return Formatted string.
      */
     private String printDecFormat(Number number) {
-        DecimalFormatSymbols otherSymbols = new DecimalFormatSymbols(Locale.US);
-        DecimalFormat decimalFormat = new DecimalFormat("#.##", otherSymbols);
         decimalFormat.setRoundingMode(RoundingMode.HALF_UP);
         return decimalFormat.format(number);
     }
